Updated with more detail in step 2 and
Had to update Step 2 as it was confusing. Got it working with the added detail. Also, the README.md in the seed has better detail. No sense in keeping track of instructions in two places. I think this makes the detail here complete, so maybe the additional details can be removed?